Change the way HeartBeat(OPTIONS) runs & PIDF bug
1. The previous heartbeat (OPTIONS message) is sent to pCSCF regardless of
whether the client (gateway) is registered successfully or not. Now if the
client is not registered, we could not get the address of sCSCF (from the
header 'Service-Route' of the 200 OK response of REGISTER request), and
hence could not send heartbeat (OPTIONS message) out, and try to
re-register the client again and again.
2. Some minor changes regarding the result code check rules in
SipCallback.
3. Bug: service node in publishing tuple is not present.

Change-Id: I8b9791360dcafe43273d77a9c3a47c914f30cc21
Signed-off-by: Thierry Luo <luoyongheng@nane.cn>