Service mapping bug fixes & Local Session Binding bug fix
1. Adding a port attribute to topmost node in introspection XML string, so
that local BusAttachment can join session with CloudServiceAgentBusObject,
since joining session should know of SessionPort in advance.
2. Local announcements do not have port information, so we should
introspect the ObjPath and/or Interface Name, to decide the well known
SessionPort, so that we can send SessionPort information to remote peer,
in order for remote BusOjbects to connect to CloudServiceAgentBusObject
through this SessionPort, otherwise, JoinSession() will fail, and so
communications between them can be established.

Change-Id: Ia5d88ab8902a33125a2f7ab8290531421a0485fe
Signed-off-by: Thierry Luo <luoyongheng@nane.cn>