Add SipEvent::GetStatus() & SipMessage:isValid() & IMSTransportSipCallback
Previously in the callback routines of IMSTransportSipCallback, we call
SipEvent::GetMessage() to create a new SipMessage, and then call
SipMessage::getResponseCode() to get the status code. Now we just call
SipEvent::GetStatus() to get the response code.

Known issue of IMSTransport & Sofia_IMS: first register is successful, and
then the OPTIONS heartbeat is not, and then a duplicate register is sent,
but after receiving 401 from IMS, subsequent register is not sent, and all
subsequent registers are not successfull with no authentication processes.

Change-Id: Ic0762891205fba6c65a1f0879b15bbcc7f3b39d0
Signed-off-by: Thierry Luo <luoyongheng@nane.cn>